Исправление обновления кнопок меню

+++ database.py
@@ -1,7 +1,9 @@
 import sqlite3
+import logging
+
 
 def set_user_locale(user_id: int, locale: str):
-    print(f"Добавление/обновление пользователя {user_id} -> {locale}") 
+    logging.info(f"Добавление/обновление пользователя {user_id} -> {locale}") 
     with sqlite3.connect("birthdays.db") as conn:
         cur = conn.cursor()
         cur.execute("""
@@ -24,31 +26,32 @@ def init_db():
     with sqlite3.connect("birthdays.db") as conn:
         cur = conn.cursor()
         cur.execute('''CREATE TABLE IF NOT EXISTS birthdays
-                       (id INTEGER PRIMARY KEY, name TEXT, date TEXT, note TEXT, contact TEXT)''')
-        conn.commit()
-    cur.execute("SELECT * FROM user_locales")
-    print(cur.fetchall())
-
-def add_birthday(name, date, note='', contact=''):
-    with sqlite3.connect("birthdays.db") as conn:
-        cur = conn.cursor()
-        cur.execute("INSERT INTO birthdays (name, date, note, contact) VALUES (?, ?, ?, ?)", (name, date, note, contact))
+                       (id INTEGER PRIMARY KEY, user_id INTEGER, name TEXT, date TEXT, note TEXT, contact TEXT)''')
+        cur.execute('''CREATE TABLE IF NOT EXISTS user_locales
+                       (user_id INTEGER PRIMARY KEY, locale TEXT)''')
         conn.commit()
 
-def get_birthdays():
+def add_birthday(user_id, name, date, note='', contact=''):
     with sqlite3.connect("birthdays.db") as conn:
         cur = conn.cursor()
-        cur.execute("SELECT name, date FROM birthdays ORDER BY date")
+        cur.execute("INSERT INTO birthdays (user_id, name, date, note, contact) VALUES (?, ?, ?, ?, ?)",
+                    (user_id, name, date, note, contact))
+        conn.commit()
+
+def get_birthdays(user_id):
+    with sqlite3.connect("birthdays.db") as conn:
+        cur = conn.cursor()
+        cur.execute("SELECT name, date FROM birthdays WHERE user_id = ? ORDER BY date", (user_id,))
         return [dict(name=row[0], date=row[1]) for row in cur.fetchall()]
 
-def delete_birthday(name):
+def delete_birthday(user_id, name):
     with sqlite3.connect("birthdays.db") as conn:
         cur = conn.cursor()
-        cur.execute("DELETE FROM birthdays WHERE name = ?", (name,))
+        cur.execute("DELETE FROM birthdays WHERE user_id = ? AND name = ?", (user_id, name))
         return cur.rowcount > 0
 
-def edit_birthday(name, new_date):
+def edit_birthday(user_id, name, new_date):
     with sqlite3.connect("birthdays.db") as conn:
         cur = conn.cursor()
-        cur.execute("UPDATE birthdays SET date = ? WHERE name = ?", (new_date, name))
+        cur.execute("UPDATE birthdays SET date = ? WHERE user_id = ? AND name = ?", (new_date, user_id, name))
         return cur.rowcount > 0
